backend: Add unit tests
This adds unit tests using the unittests package. It also documents in
the README how to run unit tests.

In order to get unit tests to work, I had to bump Flask to the latest
version. Otherwise, when executing unit tests, Flask would fail when
trying to evaluate `werkzeug.__version__`. See the following issue:

https://github.com/pallets/flask/issues/5652

I did some testing with the new Flask version and it doesn't seem to
break anything, so I think it's safe to bump.